+++ clang/lib/Format/ContinuationIndenter.cpp
@@ -407,7 +407,7 @@ bool ContinuationIndenter::mustBreak(const LineState &State) {
   if (CurrentState.BreakBeforeClosingParen && Current.is(tok::r_paren))
     return true;
   if (CurrentState.BreakBeforeClosingAngle && Current.is(TT_TemplateCloser) &&
-      Style.BreakBeforeTemplateClose) {
+      Style.BreakBeforeTemplateCloser) {
     return true;
   }
   if (Style.Language == FormatStyle::LK_ObjC &&
@@ -1378,7 +1378,7 @@ unsigned ContinuationIndenter::getNewLineColumn(const LineState &State) {
       State.Stack.size() > 1) {
     return State.Stack[State.Stack.size() - 2].LastSpace;
   }
-  if (Current.is(TT_TemplateCloser) && Style.BreakBeforeTemplateClose &&
+  if (Current.is(TT_TemplateCloser) && Style.BreakBeforeTemplateCloser &&
       State.Stack.size() > 1) {
     return State.Stack[State.Stack.size() - 2].LastSpace;
   }