+++ clang/docs/ReleaseNotes.rst
@@ -1136,7 +1136,7 @@ clang-format
 - Adds ``VariableTemplates`` option.
 - Adds support for bash globstar in ``.clang-format-ignore``.
 - Adds ``WrapNamespaceBodyWithEmptyLines`` option.
-- Adds ``BreakBeforeTemplateClose`` option.
+- Adds ``BreakBeforeTemplateCloser`` option.
 
 libclang
 --------